# swagger.yaml
openapi: 3.0.1
info:
  title: PRIN144-Final-Exam
  version: '1.0'
  description: Food Menu Management API
  contact:
    name: DANIEL ESTRELLA
    email: your.email@students.ceu.edu.ph
externalDocs:
  url: swagger.json
  description: swagger.json
servers:
  - url: https://finalsprin144.vercel.app/
    variables: {}
    description: Production
paths:
  /menu:
    get:
      summary: Returns all menu items
      description: Returns all menu items or a specific item by ID query parameter
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItemResponse'
        '404':
          description: Menu item not found
      tags:
        - Menu
    post:
      summary: Creates a new menu item
      description: Creates a new menu item record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemCreateRequest'
      responses:
        '201':
          description: Menu item created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
      tags:
        - Menu

  /menu/{id}:
    get:
      summary: Returns a menu item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemResponse'
        '404':
          description: Menu item not found
      tags:
        - Menu
    
    put:
      summary: Updates a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuItemUpdateRequest'
      responses:
        '200':
          description: Menu item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemResponse'
        '404':
          description: Menu item not found
      tags:
        - Menu

    delete:
      summary: Deletes a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Menu item deleted successfully
        '404':
          description: Menu item not found
      tags:
        - Menu

components:
  schemas:
    MenuItemCreateRequest:
      type: object
      required:
        - menu
        - category
        - price
      properties:
        menu:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        isavailable:
          type: boolean
          default: true
        preparationtimeminutes:
          type: integer

    MenuItemUpdateRequest:
      type: object
      properties:
        menu:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        isavailable:
          type: boolean
        preparationtimeminutes:
          type: integer

    MenuItemResponse:
      type: object
      properties:
        id:
          type: integer
        menu:
          type: string
        category:
          type: string
        price:
          type: number
          format: float
        isavailable:
          type: boolean
        preparationtimeminutes:
          type: integer